logic home bar chart

+++ app/src/main/java/com/example/du_an_1/funtions/MoneyFormat.java
@@ -5,7 +5,7 @@ import java.util.Locale;
 
 public class MoneyFormat {
 
-    public static String moneyFormat(int price){
+    public static String moneyFormat(double price){
         Locale localeVN = new Locale("vi", "VN");
         NumberFormat currencyFormat = NumberFormat.getCurrencyInstance(localeVN);
         return currencyFormat.format(price);
